:wrench: removed libcoiniumserv project, moved VersionInfo.cs to main project.
