+++ src/CoiniumServ/CoiniumServ.csproj
@@ -517,6 +517,7 @@
     <Compile Include="Transactions\Utils\TransactionUtils.cs" />
     <Compile Include="Utils\Numerics\BigInteger.cs" />
     <Compile Include="Utils\Numerics\BigRational.cs" />
+    <Compile Include="VersionInfo.cs" />
     <None Include="config\coins\bitcoindark.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
@@ -992,10 +993,6 @@
     <ProjectReference Include="..\..\deps\csredis\CSRedis\CSRedis.csproj">
       <Project>{d35e185e-a7e1-41e1-846c-21944f56074f}</Project>
       <Name>CSRedis</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\libCoiniumServ\libCoiniumServ.csproj">
-      <Project>{f908636c-bc75-4b93-a774-e4878f3d39af}</Project>
-      <Name>libCoiniumServ</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
