+++ src/CoiniumServ/Properties/AssemblyInfo.cs
@@ -23,7 +23,7 @@
 
 using System.Reflection;
 using System.Runtime.InteropServices;
-using libCoiniumServ.Versions;
+using CoiniumServ;
 
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
