+++ src/CoiniumServ/VersionInfo.cs
@@ -20,7 +20,7 @@
 //     license or white-label it as set out in licenses/commercial.txt.
 // 
 #endregion
-namespace libCoiniumServ.Versions
+namespace CoiniumServ
 {
     /// <summary>
     /// Supported Versions Info.
