--- src/libCoiniumServ/Properties/AssemblyInfo.cs
@@ -1,54 +1,0 @@
-﻿#region License
-// 
-//     CoiniumServ - Crypto Currency Mining Pool Server Software
-//     Copyright (C) 2013 - 2014, CoiniumServ Project - http://www.coinium.org
-//     http://www.coiniumserv.com - https://github.com/CoiniumServ/CoiniumServ
-// 
-//     This software is dual-licensed: you can redistribute it and/or modify
-//     it under the terms of the GNU General Public License as published by
-//     the Free Software Foundation, either version 3 of the License, or
-//     (at your option) any later version.
-// 
-//     This program is distributed in the hope that it will be useful,
-//     but WITHOUT ANY WARRANTY; without even the implied warranty of
-//     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//     GNU General Public License for more details.
-//    
-//     For the terms of this license, see licenses/gpl_v3.txt.
-// 
-//     Alternatively, you can license this software under a commercial
-//     license or white-label it as set out in licenses/commercial.txt.
-// 
-#endregion
-
-using System.Reflection;
-using System.Runtime.InteropServices;
-using libCoiniumServ.Versions;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("libCoiniumServ")]
-[assembly: AssemblyDescription("Next-gen crypto currency mining pool software")]
-[assembly: AssemblyCompany("Coinium.org")]
-[assembly: AssemblyProduct("libCoiniumServ - " + VersionInfo.CodeName)]
-[assembly: AssemblyCopyright("Copyright (C) 2013 - 2014, CoiniumServ project, http://www.coinium.org -  http://www.coiniumserv.com")]
-[assembly: AssemblyTrademark("libCoiniumServ")]
-[assembly: AssemblyCulture("")]
-
-#if DEBUG
-[assembly: AssemblyConfiguration("Debug")]
-#else
-    [assembly: AssemblyConfiguration("Release")]
-#endif
-
-// Setting ComVisible to false makes the types in this assembly not visible 
-// to COM components.  If you need to access a type in this assembly from 
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("0c99d5d2-11bc-4b1d-8e68-79bffd5258ed")]
-
-// Set the assembly version from VersionInfo.cs file.
-[assembly: AssemblyVersion(VersionInfo.Assembly.Version)]
